+++ R/get_captions.R
@@ -1,8 +1,9 @@
 #' Get Captions of a Video
 #' 
-#' 
-#' @param video_id id of the video; required; no default
-#' @param lang  language of the caption; required; default is english ("en")
+#' @param part  Caption resource requested. Required. Comma separated list of one or more of the 
+#' following: id, snippet. e.g., "id, snippet", "id" Default: snippet.  
+#' @param video_id ID of the video whose captions are requested. Required. No default.
+#' @param lang  Language of the caption; required; default is english ("en")
 #' @param id    comma-separated list of IDs that identify the caption resources that should be retrieved; optional; string
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
@@ -14,7 +15,7 @@
 #' get_captions(video_id="yJXTXN4xrI8")
 #' }
 
-get_captions <- function (video_id=NULL, lang="en", id=NULL, ...) {
+get_captions <- function (part="snippet", video_id=NULL, lang="en", id=NULL, ...) {
 
 	if (is.null(video_id)) stop("Must specify a video ID")
 
@@ -22,12 +23,14 @@ get_captions <- function (video_id=NULL, lang="en", id=NULL, ...) {
 
 	# Try getting captions directly
 	req <- GET(paste0("http://video.google.com/timedtext?lang=", lang, "&v=", video_id))
+	req <- content(req)
 
 	# If not try other things
-	if (length(content(req))==0) {
-		querylist = list(part="snippet", videoId = video_id, id = id)
-		req <- GET("https://www.googleapis.com/youtube/v3/captions", query = querylist, config(token = getOption("google_token")), ...)
-		stop_for_status(req)
+	if (length(req)==0) {
+		
+		querylist = list(part=part, videoId = video_id, id = id)
+		req <- tuber_GET("captions", query = querylist, ...)
+		
 		# Multiple caption tracks possible but for now harvest just the first
 		caption_id <- content(req)$items[[1]]$id
 
@@ -36,5 +39,5 @@ get_captions <- function (video_id=NULL, lang="en", id=NULL, ...) {
 		req <- NA
 	}
 
-	content(req)
+	req
 }