+++ R/get_comments.R
@@ -2,8 +2,9 @@
 #'
 #' @param video_id Data Type: Character. ID of the video. Required.
 #' @param simplify Data Type: Boolean. Default is TRUE. If TRUE, the function returns a data frame. Else a list with all the information returned.
-#' @param maxResults Data Type: Numeric. Default is 20. Takes values between 20 and 100. Optional.
-#' @param textFormat Data Type: Character. Default is "html". Only takes "html" or "plainText." Optional. 
+#' @param max_results maximum number of items that should be returned, integer, can be between 20 and 100, default is 100
+#' @param page_token specific page in the result set that should be returned, optional
+#' @param text_format Data Type: Character. Default is "html". Only takes "html" or "plainText." Optional. 
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #'  
 #' @return Nested named list. The entry \code{items} is a list of comments along with meta information. 
@@ -17,13 +18,13 @@
 #' get_comments(video_id="N708P-A45D0")
 #' }
 
-get_comments <- function (video_id=NULL, simplify=TRUE, maxResults=20, textFormat="html", ...) {
+get_comments <- function (video_id=NULL, simplify=TRUE, max_results=100, text_format="html", page_token = NULL, ...) {
 
 	if (is.null(video_id)) stop("Must specify a video ID")
-	if (maxResults < 20 | maxResults > 100) stop("maxResults only takes a value between 20 and 100")
-	if (textFormat != "html" & textFormat !="plainText") stop("Provide a legitimate value of textFormat.")
+	if (max_results < 20 | max_results > 100) stop("maxResults only takes a value between 20 and 100")
+	if (text_format != "html" & text_format !="plainText") stop("Provide a legitimate value of textFormat.")
 
-	querylist <- list(part="snippet", videoId = video_id, maxResults=maxResults, textFormat=textFormat)
+	querylist <- list(part="snippet", videoId = video_id, maxResults=max_results, textFormat=text_format)
 
 	res <- tuber_GET("commentThreads", querylist, ...)