+++ R/list_channel_sections.R
@@ -3,7 +3,8 @@
 #' Returns list of channel sections that channel id belongs to. 
 #' 
 #' @param channel_id id of the channel; required; no default
-#' @param part specify which part do you want. It can only be one of the three: contentDetails, id, snippet. Default is snippet.
+#' @param part specify which part do you want. It can only be one of the following: contentDetails, id, localizations, snippet, targeting. Default is snippet.
+#' @param hl  language that will be used for text values, optional, default is en-US. See also \code{\link{list_langs}}
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
 #' @return captions for the video from one of the first track
@@ -14,7 +15,7 @@
 #' list_channel_sections(channel_id="UCRw8bIz2wMLmfgAgWm903cA")
 #' }
 
-list_channel_sections <- function (channel_id=NULL, part="snippet", ...) {
+list_channel_sections <- function (channel_id=NULL, part="snippet", hl = NULL, ...) {
 
 	
 	if (is.null(channel_id)) stop("Must specify a channel_id")