+++ R/list_langs.R
@@ -1,6 +1,7 @@
 #' List Languages That Youtube Currently Supports
 #' 
 #' @return data.frame with 3 columns: hl (two letter abbreviation), name (of the language), etag
+#' @param hl  language that will be used for text values, optional, default is en-US. 
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
 #' @export
@@ -10,9 +11,9 @@
 #' list_langs()
 #' }
 
-list_langs <- function (...) {
+list_langs <- function (hl=NULL, ...) {
 
-	querylist <- list(part="snippet")
+	querylist <- list(part="snippet", hl = hl)
 
 	res <- tuber_GET("i18nLanguages", querylist)