+++ R/list_videos.R
@@ -1,0 +1,30 @@
+#' List Videos
+#' 
+#' @param part Required. Comma separated string including one or more of the following: contentDetails, 
+#' fileDetails, id, liveStreamingDetails, localizations, player, processingDetails, recordingDetails, 
+#' snippet, statistics, status, suggestions, topicDetails. Default: contentDetails.
+#' @param max_results maximum number of items that should be returned, integer, can be between 0 and 50, default is 50
+#' @param page_token specific page in the result set that should be returned, optional
+#' @param region_code Character. Required. Has to be a ISO 3166-1 alpha-2 code (see \url{https://www.iso.org/obp/ui/#search}).
+#' @param hl  language that will be used for text values, optional, default is en-US. See also \code{\link{list_langs}}.
+#' @param chart mostPopular
+#' @param video_category_id the video category for which the chart should be retrieved. See also \code{\link{list_videocats}}.
+#' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
+#' 
+#' @return data.frame with 5 columns: channelId, title, assignable, etag, id
+#' @export
+#' @references \url{https://developers.google.com/youtube/v3/docs/search/list}
+#' @examples
+#' \dontrun{
+#' list_videos()
+#' }
+
+list_videos <- function (chart="mostPopular", part = "contentDetails", max_results = 50, page_token=NULL, hl = NULL, region_code = NULL, video_category_id = NULL, ...) {
+
+	querylist <- list(chart="mostPopular", part=part, maxResults = max_results, pageToken = page_token, hl = hl, regionCode=region_code, videoCategoryId = video_category_id)
+
+	res <- tuber_GET("videos", querylist, ...)
+
+	res
+}
+